Add flag to also add build-dep for runtime-depends
Superflore  is correct to generate runtime dependencies when an
dependency is needed to provide headers and libraries needed to consume
an element but not to build it. Some tooling struggles to interpret the
c/c++ dependency graph when the dependencies are represented as
runtime-dependencies.

This commit adds a flag to additionally generate a build-dependency
for elements that depend on a ROS package that defines a package as
a runtime dependency.